Add a Makefile
We want to make it super simple to work with this project.

Build systems can help lower the barrier for contributions by removing
the amount of knowledge necessary to get the project running. This only
really works if the build system isn't esoteric and has a short single
command to run. `make` fits this purpose, though we should document this
setup to make it actually useful.